+++ lib/assets/javascripts/new-dashboard/pages/Home/QuotaSection/QuotaWidget.vue
@@ -14,13 +14,19 @@
 
     <div class="quota-data">
       <div class="quota-cell cell--medium">
-        <span class="text is-small is-txtSoftGrey">{{roundOneDecimal(remainingQuota)}} {{unit}}</span>
+        <span class="text is-small is-txtSoftGrey">
+          {{ formatToLocale ? getNumberInLocaleFormat(roundOneDecimal(remainingQuota)) : roundOneDecimal(remainingQuota) }} {{unit}}
+        </span>
       </div>
       <div class="quota-cell cell--medium cell--mobile">
-        <span class="text is-small is-txtSoftGrey">{{roundOneDecimal(usedQuota)}} {{unit}}</span>
+        <span class="text is-small is-txtSoftGrey">
+          {{ formatToLocale ? getNumberInLocaleFormat(roundOneDecimal(usedQuota)) : roundOneDecimal(usedQuota) }} {{unit}}
+        </span>
       </div>
       <div class="quota-cell cell--medium">
-        <span class="text is-small is-txtSoftGrey">{{roundOneDecimal(availableQuota)}} {{unit}}</span>
+        <span class="text is-small is-txtSoftGrey">
+          {{ formatToLocale ? getNumberInLocaleFormat(roundOneDecimal(availableQuota)) : roundOneDecimal(availableQuota) }} {{unit}}
+        </span>
       </div>
       <div class="quota-help cell--small">
         <a :href="helpLink" target= "_blank"><img svg-inline class="quota-image" :class="{'is-active': active}" src="../../../assets/icons/common/question-mark.svg"/></a>
@@ -38,6 +44,10 @@ export default {
     usedQuota: Number,
     unit: String,
     billingPeriod: String,
+    formatToLocale: {
+      type: Boolean,
+      default: true
+    },
     helpLink: String
   },
   data: function () {
@@ -71,6 +81,9 @@ export default {
   methods: {
     roundOneDecimal (number) {
       return Math.round(number * 10) / 10;
+    },
+    getNumberInLocaleFormat (number) {
+      return number.toLocaleString();
     },
     onMouseOver () {
       this.active = true;